Move profilerlib cpp code into headers
